define an api for openrisk
