+++ .gitignore
@@ -7,7 +7,7 @@
 *.dll
 *.so
 *.dylib
-openrisk
+
 
 # Test binary, built with `go test -c`
 *.test
