+++ openrisk/types.go
@@ -1,4 +1,19 @@
-package main
+package openrisk
+
+import gogpt "github.com/sashabaranov/go-gpt3"
+
+type OpenRisk struct {
+	client gogpt.Client
+}
+
+type Options struct {
+	ApiKey string
+}
+
+type NucleiScan struct {
+	Issues string
+	Score  string
+}
 
 // The JSON types do not define the entire structure of the JSON object that is exported by Nuclei, but rather only
 // the fields that are used in the issue list generation. This reduces the amount of data that is parsed from the JSON
